+++ Honours/Assets/Scripts/Player/Wolf.cs
@@ -355,7 +355,7 @@ public class Wolf : MonoBehaviour, IDamageable
         }
     }
 
-    public void Damage(float damage)
+    public void Damage(int damage)
     {
         if (isDead) { return; }
         // Current health is decreased by the damage received
